split thin bridge surfaces
